+++ src/ebarimt_pos_sdk/resources/api/district/schema.py
@@ -1,0 +1,14 @@
+from ...base_model import BaseEbarimtModel
+
+
+class BranchInfo(BaseEbarimtModel):
+    branch_code: str
+    branch_name: str
+    sub_branch_code: str
+    sub_branch_name: str
+
+
+class GetDistrictCodeResponse(BaseEbarimtModel):
+    msg: str
+    status: int
+    data: list[BranchInfo]